waveform 'old_texture' storage for time limit early exit condition (ugh); better handling of timeline bounds that fixes mysterious audio hang bug, hopefully; different look for active tracks

+++ src/timeview.c
@@ -42,13 +42,13 @@ int32_t timeview_rightmost_pos(TimeView *tv)
 static void timeview_rectify_scroll(TimeView *tv)
 {
     if (tv->restrict_view) {
-	int32_t overshoot_right = 0;
+	int64_t overshoot_right = 0;
 	/* Like crossing the street */
 	if (tv->offset_left_sframes < tv->view_min) {
 	    tv->offset_left_sframes = tv->view_min;
 	    fprintf(stderr, "Offset left dropped below min\n");
 	}
-	if ((overshoot_right = timeview_rightmost_pos(tv) - tv->view_max) > 0) {
+	if ((overshoot_right = (int64_t)timeview_rightmost_pos(tv) - tv->view_max) > 0) {
 	    tv->offset_left_sframes -= overshoot_right;
 	    fprintf(stderr, "Offset right dropped below min\n");
 	}
